+++ lib/ibandit/local_details_cleaner.rb
@@ -1,7 +1,7 @@
 module Ibandit
   module LocalDetailsCleaner
-    SUPPORTED_COUNTRY_CODES = %w(AT BE BG CY DE DK EE ES FI FR GB GR HU IE IS IT
-                                 LT LU LV MC MT NL NO PL PT RO SE SI SK
+    SUPPORTED_COUNTRY_CODES = %w(AT BE BG CY CZ DE DK EE ES FI FR GB GR HR HU IE
+                                 IS IT LT LU LV MC MT NL NO PL PT RO SE SI SK
                                  SM).freeze
 
     def self.clean(local_details)
@@ -30,7 +30,7 @@ module Ibandit
       case country_code
       when 'AT', 'CY', 'DE', 'FI', 'LT', 'LU', 'LV', 'NL', 'RO', 'SI', 'SK'
         %i(bank_code account_number)
-      when 'BE', 'DK', 'EE', 'ES', 'HU', 'IS', 'NO', 'PL', 'SE'
+      when 'BE', 'CZ', 'DK', 'EE', 'ES', 'HR', 'HU', 'IS', 'NO', 'PL', 'SE'
         %i(account_number)
       when 'GB', 'IE', 'MT'
         if Ibandit.bic_finder.nil? then %i(bank_code branch_code account_number)
@@ -91,6 +91,26 @@ module Ibandit
       {
         bank_code:      bank_code,
         branch_code:    branch_code,
+        account_number: account_number
+      }
+    end
+
+    def self.clean_cz_details(local_details)
+      #   The SWIFT definition of a Czech IBAN includes both the account
+      #   number prefix and the account number. This method therefore supports
+      #   passing those fields concatenated.
+      account_number =
+        if local_details.include?(:account_number_prefix)
+          [
+            local_details[:account_number_prefix].rjust(6, '0'),
+            local_details[:account_number].rjust(10, '0')
+          ].join
+        else
+          local_details[:account_number].tr('-', '').rjust(16, '0')
+        end
+
+      {
+        bank_code:      local_details[:bank_code],
         account_number: account_number
       }
     end
@@ -231,6 +251,20 @@ module Ibandit
       # Greek IBANs construction is idiosyncratic to the individual banks, and
       # no central specification is published.
       local_details
+    end
+
+    def self.clean_hr_details(local_details)
+      # This method supports being passed the component IBAN parts, as defined
+      # by SWIFT, or a single traditional-format string split by a '-'.
+      return local_details if local_details[:bank_code]
+      return local_details unless local_details[:account_number].include?('-')
+
+      bank_code, account_number = local_details[:account_number].split('-', 2)
+
+      {
+        bank_code:      bank_code,
+        account_number: account_number
+      }
     end
 
     def self.clean_hu_details(local_details)
@@ -423,23 +457,8 @@ module Ibandit
     end
 
     def self.clean_sk_details(local_details)
-      #   The SWIFT definition of a Slovakian IBAN includes both the account
-      #   number prefix and the account number. This method therefore supports
-      #   passing those fields concatenated.
-      account_number =
-        if local_details.include?(:account_number_prefix)
-          [
-            local_details[:account_number_prefix].rjust(6, '0'),
-            local_details[:account_number].rjust(10, '0')
-          ].join
-        else
-          local_details[:account_number].tr('-', '').rjust(16, '0')
-        end
-
-      {
-        bank_code:      local_details[:bank_code],
-        account_number: account_number
-      }
+      # Slovakia uses the same local details method as the Czech Republic
+      clean_cz_details(local_details)
     end
 
     def self.clean_sm_details(local_details)